Rewrite put/delete favorite actions
Rewriting to have proper routes like GitHub starring API.
The routes are now like that :

* PUT /user/favorites/{game}
* DELETE /user/favorites/{game}

If the user try to favorite an already favorited game, nothing is done
but the response is no different than a succeeded request. The user
wanted it favorite, so it still the case.
Furthermore, the DELETE action is idempotent so if the favorite
doesn't exist, it doesn't matter, the user wanted it gone, and it
was already the case.